the file app.js was rename for MovieApp.js
